import excel

+++ database/migrations/2024_12_07_075733_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string("product_id");
             $table->string("name");
-            $table->integer("price");
+            $table->float("price");
             $table->timestamps();
         });
     }
